Integrated move into map editor

+++ src/main/java/edu/wpi/teamA/entities/MapEntity.java
@@ -184,7 +184,11 @@ public class MapEntity {
 
   public LocationName getLocationName(int nodeID) {
     Move move = databaseRepo.getMoveForNode(nodeID);
-    return databaseRepo.getLocName(move.getLongName());
+    if (move != null) {
+      return databaseRepo.getLocName(move.getLongName());
+    } else {
+      return databaseRepo.getLocName("Empty Node");
+    }
   }
 
   public void determineAddAction(String level, Node node, LocationName locName, Move move) {
@@ -205,15 +209,23 @@ public class MapEntity {
       determineEdgeMap(level).remove(key);
     }
     databaseRepo.deleteNode(databaseRepo.getNode(nodeID));
-    databaseRepo.deleteLocName(
-        databaseRepo.getLocName(databaseRepo.getMoveForNode(nodeID).getLongName()));
-    databaseRepo.deleteMove(databaseRepo.getMoveForNode(nodeID));
+    Move move = databaseRepo.getMoveForNode(nodeID);
+    if (move != null) {
+      databaseRepo.deleteLocName(databaseRepo.getLocName(move.getLongName()));
+      databaseRepo.deleteMove(databaseRepo.getMoveForNode(nodeID));
+    }
     determineNodeMap(level).remove(nodeID);
   }
 
   public void determineModifyAction(String level, Node node, LocationName locName, Move move) {
     databaseRepo.updateNode(node);
-    databaseRepo.updateLocName(locName);
+    String longname = locName.getLongName();
+    if (!longname.equals("Empty Node")) {
+      databaseRepo.updateLocName(locName);
+    } else {
+      databaseRepo.addLocName(locName);
+    }
+
     databaseRepo.addMove(move);
     determineNodeMap(level).put(node.getNodeID(), node);
   }
@@ -299,6 +311,13 @@ public class MapEntity {
       } else if (DAOimp.equals("Edge")) {
         databaseRepo.exportData(selectedDirectory.getPath(), DAOimp);
       }
+    }
+  }
+
+  public void submitNewMoves(Move move, LocationName locationName) {
+    databaseRepo.addMove(move);
+    if (!databaseRepo.getLocNameMap().containsKey(locationName.getLongName())) {
+      databaseRepo.addLocName(locationName);
     }
   }
